KalshiBot/trader/requirements.txt
# Trader (Fast Brain) Dependencies
websocket-client>=1.7.0
cryptography>=41.0.0
requests>=2.31.0
python-dotenv>=1.0.0
scipy>=1.10.0
